firmware/requirements.txt
argparse===1.4.0
Jetson.GPIO===2.0.4